# backend/openapi.yaml
swagger: "2.0"
info:
  title: PlayGrade API
  description: API for the PlayGrade entertainment discussion platform.
  version: 1.0.0
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: Enter a JWT as `Bearer <token>`.
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: List and filter posts
      parameters:
        - $ref: "#/parameters/OptionalAuthorization"
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: limit
          type: integer
          default: 10
        - in: query
          name: posterId
          type: integer
        - in: query
          name: categories
          type: string
          description: Comma-separated display category names.
        - in: query
          name: users
          type: string
          enum: [All Users, Followed Users]
        - in: query
          name: ageRange
          type: string
          enum: [Today, Week, Month, Year, All]
        - in: query
          name: sortBy
          type: string
          enum: [Newest, Most Liked, Most Comments]
        - in: query
          name: searchQuery
          type: string
      responses:
        200:
          description: Paginated posts
    post:
      tags: [Posts]
      summary: Create a post
      consumes:
        - multipart/form-data
      security:
        - Bearer: []
      parameters:
        - in: formData
          name: title
          required: true
          type: string
          maxLength: 100
        - in: formData
          name: body
          required: true
          type: string
          maxLength: 300
        - in: formData
          name: category
          required: true
          type: string
          enum: [🎮 Games, 🎥 Film/TV, 🎵 Music]
        - in: formData
          name: image
          required: true
          type: file
      responses:
        201:
          description: Post created
        400:
          description: Invalid post
  /posts/{post_id}:
    parameters:
      - $ref: "#/parameters/PostId"
    get:
      tags: [Posts]
      summary: Get a post and its replies
      parameters:
        - $ref: "#/parameters/OptionalAuthorization"
      responses:
        200:
          description: Post details and replies
        404:
          description: Post not found
    delete:
      tags: [Posts]
      summary: Delete a post
      security:
        - Bearer: []
      responses:
        200:
          description: Post deleted
        403:
          description: Not permitted
        404:
          description: Post not found
...
  /uploads/{filename}:
    get:
      tags: [Files]
      summary: Get an uploaded image
      produces:
        - image/png
        - image/jpeg
        - image/gif
      parameters:
        - in: path
          name: filename
          required: true
          type: string
      responses:
        200:
          description: Uploaded image
          schema:
            type: file
        404:
          description: File not found
parameters:
...
  PostId:
    in: path
    name: post_id
    required: true
    type: integer
...
  OptionalAuthorization:
    in: header
    name: Authorization
    required: false
    type: string
    description: Optional `Bearer <token>` value.
...